Apache 2 licensing and attribution
Hi Joshua! I was scrolling Github looking for Hue driver, and while scrolling, I noticed that you're using some of my MagicHome code in your project verbatim to my original writing in your project. 

This is fine—and it's completely fair game to use—but it's licensed under Apache 2 and requires attribution. Specifically, the imports, logDebug preferences and method,  deviceIP preferences, parts of your commands ( '// 0 - 255' ), and parse, are verbatim to my original code.

It looks like this is the only file affected, so I've made a PR with attribution added in hopes that you'll work with me in honoring the Apache 2 license listed in my drivers.